requirements.txt
streamlit
pandas
pillow
pytesseract
matplotlib
openpyxl
